chore: move gl context to separate package
