+++ src/main/java/dev/silenium/compose/gl/context/EGLContext.kt
@@ -1,4 +1,4 @@
-package dev.silenium.compose.gl.fbo
+package dev.silenium.compose.gl.context
 
 import org.lwjgl.egl.EGL
 import org.lwjgl.egl.EGL15.*
